+++ tests/models/independent/test_Sample.py
@@ -2,7 +2,10 @@
 
 from collections import Mapping
 import copy
+import pickle
+import tempfile
 import os
+
 
 import mock
 import numpy as np
@@ -484,6 +487,19 @@ class SampleConstructorTests:
         assert {} == diff
         extra = [v for k, v in obs_meta.items() if k not in exp_meta]
         assert not any(extra)
+
+
+class SampleSerializationTests:
+    """ Basic tests of Sample serialization with pickle module. """
+
+    def test_pickle_roundtrip(self):
+        """ Test whether pickle roundtrip produces a comparable object """
+        s = Sample({SAMPLE_NAME_COLNAME: "testsample"})
+        _buffer = tempfile.TemporaryFile()
+        pickle.dump(s, _buffer)
+        _buffer.seek(0)
+        new_s = pickle.load(_buffer)
+        assert s == new_s
 
 
 def _get_prj(conf_file, data, proj_type):